[Savestates] Start rewriting player motion
