+++ src/main/java/com/minecrafttas/tasmod/networking/TASmodBufferBuilder.java
@@ -10,61 +10,61 @@ import java.nio.ByteBuffer;
 import com.minecrafttas.mctcommon.networking.ByteBufferBuilder;
 import com.minecrafttas.mctcommon.networking.interfaces.PacketID;
 import com.minecrafttas.tasmod.playback.PlaybackControllerClient.TASstate;
-import com.minecrafttas.tasmod.savestates.handlers.SavestatePlayerHandler.MotionData;
+import com.minecrafttas.tasmod.savestates.storage.SavestateMotionStorage.MotionData;
 import com.minecrafttas.tasmod.tickratechanger.TickrateChangerServer.TickratePauseState;
 
 import net.minecraft.nbt.CompressedStreamTools;
 import net.minecraft.nbt.NBTSizeTracker;
 import net.minecraft.nbt.NBTTagCompound;
 
-public class TASmodBufferBuilder extends ByteBufferBuilder{
+public class TASmodBufferBuilder extends ByteBufferBuilder {
 
 	public TASmodBufferBuilder(int id) {
 		super(id);
 	}
-	
+
 	public TASmodBufferBuilder(PacketID packet) {
 		super(packet);
 	}
-	
+
 	public TASmodBufferBuilder(ByteBuffer buf) {
 		super(buf);
 	}
-	
+
 	public TASmodBufferBuilder writeTASState(TASstate state) {
-		this.writeShort((short)state.ordinal());
+		this.writeShort((short) state.ordinal());
 		return this;
 	}
-	
+
 	public TASmodBufferBuilder writeNBTTagCompound(NBTTagCompound compound) {
-		
+
 		ByteArrayOutputStream out = new ByteArrayOutputStream();
-		
+
 		DataOutputStream dataout = new DataOutputStream(out);
-		
+
 		try {
 			CompressedStreamTools.write(compound, dataout);
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-		
+
 		this.writeByteArray(out.toByteArray());
-		
+
 		try {
 			out.close();
 			dataout.close();
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-		
+
 		return this;
 	}
-	
+
 	public TASmodBufferBuilder writeTickratePauseState(TickratePauseState state) {
 		writeShort((short) state.ordinal());
 		return this;
 	}
-	
+
 	public TASmodBufferBuilder writeMotionData(MotionData data) {
 		writeDouble(data.getClientX());
 		writeDouble(data.getClientY());
@@ -76,28 +76,28 @@ public class TASmodBufferBuilder extends ByteBufferBuilder{
 		writeBoolean(data.isSprinting());
 		return this;
 	}
-	
+
 	public static TASstate readTASState(ByteBuffer buf) {
 		return TASstate.values()[buf.getShort()];
 	}
-	
+
 	public static NBTTagCompound readNBTTagCompound(ByteBuffer buf) throws IOException {
 		ByteArrayInputStream input = new ByteArrayInputStream(readByteArray(buf));
-		
+
 		DataInputStream datain = new DataInputStream(input);
-		
+
 		NBTTagCompound compound = CompressedStreamTools.read(datain, NBTSizeTracker.INFINITE);
-		
+
 		input.close();
 		datain.close();
-		
+
 		return compound;
 	}
-	
+
 	public static TickratePauseState readTickratePauseState(ByteBuffer buf) {
 		return TickratePauseState.values()[buf.getShort()];
 	}
-	
+
 	public static MotionData readMotionData(ByteBuffer buf) {
 		double x = TASmodBufferBuilder.readDouble(buf);
 		double y = TASmodBufferBuilder.readDouble(buf);
@@ -107,8 +107,8 @@ public class TASmodBufferBuilder extends ByteBufferBuilder{
 		float rz = TASmodBufferBuilder.readFloat(buf);
 		float jumpMovementVector = TASmodBufferBuilder.readFloat(buf);
 		boolean sprinting = TASmodBufferBuilder.readBoolean(buf);
-		
+
 		return new MotionData(x, y, z, rx, ry, rz, sprinting, jumpMovementVector);
 	}
-	
+
 }
